best_model/requirements.txt
torch>=2.0
torchvision>=0.15
pillow
streamlit>=1.33
pandas>=1.5
